Candidates

See merge request my-awesome-group2730631/my_awesome_project!6

+++ project_hr/src/components/Table/SelectActions/useSelectActions.tsx
@@ -1,28 +1,40 @@
 import { useCallback, useState } from 'react';
 import type { SelectChangeEvent } from '@mui/material';
 
+import { useDeleteCandidates } from 'api/candidate/deleteCandidate/useDeleteCandidates';
 import { useDeleteJobs } from 'api/job/deleteJob/useDeleteJobs';
 
 import type { UseCheckboxRowProps } from '../CheckboxRow/CheckboxRow.types';
+import type { DataCategory } from '../Table';
 
-import type { SelectActionsValue } from './SelectActions.types';
+import type { SelectActionsValue } from './actions';
 
-export const useSelectActions = (checkboxRow: UseCheckboxRowProps) => {
+export const useSelectActions = (
+  dataCategory: DataCategory,
+  checkboxRow: UseCheckboxRowProps,
+) => {
   const [value, setValue] = useState<SelectActionsValue>('Actions');
-  const { markedRows, unmarkAllRows } = checkboxRow;
 
-  const { handleDeleteJobs } = useDeleteJobs(markedRows);
+  const { handleDeleteCandidates } = useDeleteCandidates(
+    checkboxRow.markedRows,
+  );
+  const { handleDeleteJobs } = useDeleteJobs(checkboxRow.markedRows);
 
   const handleChange = useCallback(
     (event: SelectChangeEvent) => {
       const item = event.target.value as SelectActionsValue;
 
-      if (item === 'Delete') {
+      if (item === 'Delete' && dataCategory === 'candidates') {
+        handleDeleteCandidates();
+        checkboxRow.unmarkAllRows();
+      }
+
+      if (item === 'Delete' && dataCategory === 'jobs') {
         handleDeleteJobs();
-        unmarkAllRows();
+        checkboxRow.unmarkAllRows();
       }
     },
-    [handleDeleteJobs, unmarkAllRows],
+    [handleDeleteCandidates, handleDeleteJobs, dataCategory, checkboxRow],
   );
 
   return { value, handleChange };
